+++ tests/testthat/test-joins.r
@@ -558,3 +558,11 @@ test_that("joins work with factors of different levels (#1712)", {
   res2 <- left_join(d1, d2, by="Species")
   expect_equal(res1, res2)
 })
+
+test_that("anti joins give correct result when by variable is a factor (#1571)", {
+  big <- data.frame(letter = rep(c('a', 'b'), each = 2), number = 1:2)
+  small <- data.frame(letter = 'b')
+  expect_message( aj_result <- anti_join(big, small), 'Joining by: "letter"' )
+  expect_equal( aj_result$number, 1:2)
+  expect_equal( aj_result$letter, factor(c("a", "a"), levels = c("a", "b")) )
+})